requirements.txt
fitparse==1.2.0
geopy==2.4.1
gpxpy==1.6.2
olefile==0.47
Pillow==10.1.0
requests==2.31.0
attrs==23.1.0
streamlit==1.29.0
watchdog==3.0.0
streamlit-folium==0.17.2
flickrapi==2.4.0
eyed3==0.9.7
Flask==3.0.0
geojson==3.1.0

pytest==7.4.3
